requirements.txt
absl-py==0.9.0
addict @ file:///home/rick/Dropbox/python_projects/data_science/Kaggle/hpa2/input/mmdetection-v280/mmdetection/src/addict-2.4.0-py3-none-any.whl
aiohttp==3.7.4.post0
albumentations==0.4.5
antlr4-python3-runtime==4.8
appdirs==1.4.3
astroid==2.4.1
async-timeout==3.0.1
attrs==19.3.0
backcall==0.1.0
bleach==3.1.5
cachetools==4.1.0
catalyst==20.11
certifi==2020.4.5.1
chardet==3.0.4
charset-normalizer==2.0.3
click==8.0.1
comet-git-pure==0.19.16
comet-ml==3.1.6
configobj==5.0.6
configparser==5.0.2
cycler==0.10.0
Cython==0.29.17
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
deprecation==2.1.0
docker-pycreds==0.4.0
easydict==1.9
effdet==0.2.4
efficientnet-pytorch==0.7.0
ensemble-boxes==1.0.6
entrypoints==0.3
everett==1.0.2
fastcore==1.3.25
fsspec==2021.7.0
future==0.18.2
gitdb==4.0.5
GitPython==3.1.11
google-auth==1.14.2
google-auth-oauthlib==0.4.1
grpcio==1.28.1
hydra-core==1.0.4
idna==2.9
imagecodecs==2020.2.18
ImageHash==4.2.0
imageio==2.8.0
imgaug==0.2.6
importlib-metadata==1.6.0
importlib-resources==3.0.0
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.17.0
Jinja2==2.11.2
joblib==0.14.1
json5==0.9.5
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==2.2.9
jupyterlab-server==1.2.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
llvmlite==0.36.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.4.2
mccabe==0.6.1
mistune==0.8.4
mmcv-full==1.2.6
mmdet @ file:///home/rick/Dropbox/python_projects/data_science/Kaggle/hpa2/input/mmdetection-v280/mmdetection/src/mmdet-2.8.0
mmpycocotools @ file:///home/rick/Dropbox/python_projects/data_science/Kaggle/hpa2/input/mmdetection-v280/mmdetection/src/mmpycocotools-12.0.3
multidict==5.1.0
nbconvert==5.6.1
nbformat==5.0.6
netifaces==0.10.9
networkx==2.4
notebook==6.0.3
numba==0.53.1
numpy==1.20.3
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
objdetecteval @ git+https://github.com/alexhock/object-detection-metrics@4ba0cfa7677f8f0b65776567b6ccb7b45ac3b1e0
omegaconf==2.0.5
opencv-python==4.5.1.48
opencv-python-headless==4.2.0.34
packaging==20.3
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
plotly==4.12.0
pooch==1.1.0
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==3.0.5
protobuf==3.17.1
psutil==5.8.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pyDeprecate==0.3.1
Pygments==2.6.1
pylint==2.5.2
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytorch-lightning==1.4.0
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.3
QtPy==1.9.0
requests==2.26.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.0
scikit-image==0.17.1
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.11.1
Send2Trash==1.5.0
sentry-sdk==1.1.0
shortuuid==1.0.1
six==1.14.0
smmap==3.0.4
subprocess32==3.5.4
tensorboard==2.4.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.6.0.post3
tensorboardX==2.1
terminado==0.8.3
terminaltables==3.1.0
testpath==0.4.4
thop==0.0.31.post2005241907
tifffile==2020.5.7
timm==0.4.12
toml==0.10.0
torch==1.9.0
torchaudio==0.9.0
torchmetrics==0.4.1
torchvision==0.10.0
tornado==6.0.4
tqdm==4.45.0
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.26.6
vision-transformer-pytorch==1.0.2
wandb==0.11.0
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
wurlitzer==2.0.0
yapf @ file:///home/rick/Dropbox/python_projects/data_science/Kaggle/hpa2/input/mmdetection-v280/mmdetection/src/yapf-0.30.0-py2.py3-none-any.whl
yarl==1.6.3
zipp==3.1.0
